backend/requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.16.0
mysqlclient==2.2.7
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
django-cors-headers
python-dotenv